+++ src/libs/karm-ui/input.cpp
@@ -281,7 +281,7 @@ struct Input : public View<Input> {
         g.save();
 
         auto m = mesure();
-        auto baseline = bound().topStart() + m.baseline.cast<isize>();
+        auto baseline = bound().topStart() + m.baseline;
 
         if (_style.color) {
             g.fillStyle(*_style.color);
